mahalaxmi-backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.3
drf-spectacular==0.26.5
dj-database-url==2.1.0
reportlab==4.5.0
cloudinary==1.40.0
